# backend/docs/swagger.yaml
basePath: /
definitions:
...
  tickets.CreateTicketRequest:
    properties:
      description:
        type: string
      subject:
        type: string
    required:
    - description
    - subject
    type: object
  tickets.Ticket:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      installer_id:
        description: Can be nil initially
        type: string
      installer_name:
        type: string
      status:
        $ref: '#/definitions/tickets.TicketStatus'
      subject:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      user_name:
        description: Joined fields
        type: string
    type: object
  tickets.TicketStatus:
    enum:
    - OPEN
    - IN_PROGRESS
    - RESOLVED
    - CLOSED
    type: string
    x-enum-varnames:
    - StatusOpen
    - StatusInProgress
    - StatusResolved
    - StatusClosed
  tickets.UpdateTicketRequest:
    properties:
      status:
        $ref: '#/definitions/tickets.TicketStatus'
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for the Solar Energy Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Energy Management System API
  version: "1.0"
paths:
...
  /installer/tickets:
    get:
      consumes:
      - application/json
      description: Get tickets assigned to current installer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tickets.Ticket'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get installer tickets
      tags:
      - Tickets
  /installer/tickets/{id}/status:
    put:
      consumes:
      - application/json
      description: Update status of a ticket
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/tickets.UpdateTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update ticket status
      tags:
      - Tickets
...
  /user/tickets:
    get:
      consumes:
      - application/json
      description: Get current user's tickets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tickets.Ticket'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user tickets
      tags:
      - Tickets
    post:
      consumes:
      - application/json
      description: Create a new support ticket
      parameters:
      - description: Ticket creation data
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/tickets.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tickets.Ticket'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a ticket
      tags:
      - Tickets
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"